requirements.txt
streamlit==0.88.0
tensorflow-cpu
Pillow==8.2.0
numpy==1.19.5
